Move OutputPortListenerInterface into its own file and the drake::systems::detail namespace.
