+++ drake/systems/framework/CMakeLists.txt
@@ -9,6 +9,7 @@ set(sources
   examples/spring_mass_system.cc
   leaf_context.cc
   leaf_system.cc
+  output_port_listener_interface.cc
   primitives/adder.cc
   primitives/constant_value_source.cc
   primitives/constant_vector_source.cc
@@ -45,6 +46,7 @@ set(installed_headers
   diagram_context.h
   leaf_context.h
   leaf_system.h
+  output_port_listener_interface.h
   state.h
   state_subvector.h
   state_supervector.h
