+++ drake/systems/framework/output_port_listener_interface.cc
@@ -1,0 +1,11 @@
+#include "drake/systems/framework/output_port_listener_interface.h"
+
+namespace drake {
+namespace systems {
+namespace detail {
+
+OutputPortListenerInterface::~OutputPortListenerInterface() {}
+
+}  // namespace detail
+}  // namespace systems
+}  // namespace drake
